django_core/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-jazzmin==2.4.8
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
django-filter==21.1
django-lifecycle==1.0.0
django-import-export==2.7.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pgsql==1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
celery==5.1.2
redis==3.5.3
django-constance==2.9.0
django-redis==5.2.0
drf-yasg==1.20.0
pillow==9.2.0
numpy==1.23.4
opencv-python==4.6.0.66
cvzone==1.5.6